+++ sdk/keyvault/arm-keyvault/src/classic/index.ts
@@ -1,0 +1,14 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+export type { KeysOperations } from "./keys/index.js";
+export type { ManagedHsmKeysOperations } from "./managedHsmKeys/index.js";
+export type { ManagedHsmsOperations } from "./managedHsms/index.js";
+export type { MhsmPrivateEndpointConnectionsOperations } from "./mhsmPrivateEndpointConnections/index.js";
+export type { MhsmPrivateLinkResourcesOperations } from "./mhsmPrivateLinkResources/index.js";
+export type { MhsmRegionsOperations } from "./mhsmRegions/index.js";
+export type { OperationsOperations } from "./operations/index.js";
+export type { PrivateEndpointConnectionsOperations } from "./privateEndpointConnections/index.js";
+export type { PrivateLinkResourcesOperations } from "./privateLinkResources/index.js";
+export type { SecretsOperations } from "./secrets/index.js";
+export type { VaultsOperations } from "./vaults/index.js";