+++ sdk/kubernetesconfiguration/arm-kubernetesconfiguration-privatelinkscopes/src/api/privateLinkScopes/index.ts
@@ -1,0 +1,19 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+export {
+  list,
+  listByResourceGroup,
+  $delete,
+  updateTags,
+  createOrUpdate,
+  get,
+} from "./operations.js";
+export type {
+  PrivateLinkScopesListOptionalParams,
+  PrivateLinkScopesListByResourceGroupOptionalParams,
+  PrivateLinkScopesDeleteOptionalParams,
+  PrivateLinkScopesUpdateTagsOptionalParams,
+  PrivateLinkScopesCreateOrUpdateOptionalParams,
+  PrivateLinkScopesGetOptionalParams,
+} from "./options.js";